Fix: Angular implmentation updated 25.1.3
