+++ Angular/src/app/app.component.scss
@@ -1,4 +1,13 @@
-.default-style {
-  margin: 50px;
-  width: 90vw;
+::ng-deep .cell-selected {
+  background-color: lightgreen;
+}
+
+::ng-deep .dx-datagrid {
+  user-select: none;
+}
+
+::ng-deep .selected-data {
+  margin-top: 10px;
+  overflow-y: scroll;
+  max-height: 200px;
 }
