--- Angular/src/app/orig_app.component.html
@@ -1,19 +1,0 @@
-<dx-data-grid
-  id="gridContainer"
-  [dataSource]="customers"
-  keyExpr="ID"
-  [columns]="['CompanyName', 'City', 'State', 'Phone', 'Fax']"
-  [showBorders]="true"
-  (onCellHoverChanged) = "onCellHoverChanged($event)"
-  (onCellClick) = "onCellClick($event)"
-  (onCellPrepared) = "onCellPrepared($event)"
-  (onContentReady) = "onContentReady($event)"
->
-</dx-data-grid>
-<div class="selected-data">
-  <h3 class="caption">Selected Cells:</h3>
-
-  <ul>
-    <li *ngFor='let info of data'>{{ getCellValue(info) }}</li>
-  </ul>
-</div>
